% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BH_method.R
\name{BH_method}
\alias{BH_method}
\title{Function for selecting variable under the tolerance level of FDR (False Discovery Rate)}
\usage{
BH_method(tol, selection_total)
}
\arguments{
\item{tol}{tolerance level vector or single element}

\item{selection_total}{selection counts vector of size 'p'}
}
\description{
Using the Idea of Benjamini and Hochberg method.
}
