% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evs.R
\name{evs}
\alias{evs}
\title{Big function for Ensemble Variable Selection using random selection}
\usage{
evs(
  x,
  y,
  B1,
  B2,
  d,
  model = "gaussian",
  tol = c(0.1, 0.05, 0.01),
  parallel = FALSE,
  cl = NULL,
  n_cand = 10,
  control_opt = list(strata = NULL, init = NULL, offset = NULL, eps = 1e-09, maxiter =
    20, weights = NULL, method = "efron", rownames = NULL)
)
}
\arguments{
\item{x}{data matrix}

\item{y}{response vector or 'Surv' class object for CoxPH model}

\item{B1}{Total number of ensemble iteration}

\item{B2}{Total number of iterations for weak learners}

\item{d}{Random selection dimension}

\item{model}{Either 'gaussian', 'logistic', 'poisson' or 'coxph'}

\item{tol}{tolerance level vector, default setting is 0.1, 0.05, 0.01}

\item{parallel}{Boolean variable for parallel computation}

\item{cl}{'makeCluster()' object}

\item{n_cand}{number of candidate beta coefficient for warm start}

\item{control_opt}{list variable for hyper-parameters in coxph.fit. Similar to coxph.control()}
}
\description{
Big function for Ensemble Variable Selection using random selection
}
