% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_exact_pval.R
\name{get_exact_pval}
\alias{get_exact_pval}
\title{Function for p value of binomial distribution (One(right)-sided)}
\usage{
get_exact_pval(value, B1, d, p)
}
\arguments{
\item{value}{selected counts}

\item{B1}{Iteration for Ensemble}

\item{d}{Dimension for random selection}

\item{p}{Number of covariates}
}
\description{
Function for p value of binomial distribution (One(right)-sided)
}
