% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_coxph_ploglik.R
\name{my_coxph_ploglik}
\alias{my_coxph_ploglik}
\title{Partial log-likelihood for coxph model}
\usage{
my_coxph_ploglik(beta, x, y)
}
\arguments{
\item{beta}{beta coefficients}

\item{x}{data matrix}

\item{y}{'Surv' class variable}
}
\description{
Partial log-likelihood for coxph model
}
