% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_logit_loglik.R
\name{my_logit_loglik}
\alias{my_logit_loglik}
\title{Log-likelihood for logistic regression model}
\usage{
my_logit_loglik(beta, x, y)
}
\arguments{
\item{beta}{beta coefficients}

\item{x}{data matrix}

\item{y}{binary response variable}
}
\description{
Log-likelihood for logistic regression model
}
