% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_mse_index.R
\name{my_mse_index}
\alias{my_mse_index}
\title{-1 * MSE of linear regression model}
\usage{
my_mse_index(index, x, y)
}
\arguments{
\item{index}{index vector}

\item{x}{data matrix}

\item{y}{continuous response vector}
}
\description{
-1 * MSE of linear regression model
}
