% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_pois_loglik.R
\name{my_pois_loglik}
\alias{my_pois_loglik}
\title{Log-likelihood for poisson regression model}
\usage{
my_pois_loglik(beta, x, y)
}
\arguments{
\item{beta}{beta coefficients}

\item{x}{data matrix}

\item{y}{non-negative response variable (counts)}
}
\description{
Log-likelihood for poisson regression model
}
