% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EVS.R
\name{sample.cont}
\alias{sample.cont}
\title{Samples selected quantile of continuous variable}
\usage{
sample.cont(y, qntl)
}
\arguments{
\item{y}{numeric vector}

\item{qntl}{upper and lower quantile to sample}
}
\description{
Samples selected quantile of continuous variable
}
