% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/temp_parallel.R
\name{temp_parallel}
\alias{temp_parallel}
\title{Function for parallel computation}
\usage{
temp_parallel(x, y, b1, B2, n, p, d, model)
}
\arguments{
\item{x}{data matrix}

\item{y}{Response variable in vector or 'Surv' class object for CoxPH model}

\item{b1}{Iteration for Ensemble}

\item{B2}{Total number of iteration for selecting weak learners}

\item{n}{sample size}

\item{p}{Number of covariates}

\item{d}{Dimension for random selection}

\item{model}{Either 'gaussian', 'logistic', 'poisson' or 'coxph'}
}
\description{
Function for parallel computation
}
